Inject client into the application service wrapper
