+++ internal/s3/client.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/minio/minio-go/v7"
-	"github.com/minio/minio-go/v7/pkg/credentials"
 )
 
 type Statuses = string
@@ -26,37 +25,6 @@ type IClient interface {
 	CreateBucket(ctx context.Context, bucketName string) error
 	BucketExists(ctx context.Context, bucketName string) (bool, error)
 	FileExists(ctx context.Context, filename, bucketName string) (bool, error)
-}
-
-type S3Config struct {
-	Host                 string `env:"HOST"`
-	Port                 int    `env:"PORT"`
-	AccessKey            string `env:"ACCESS_KEY"`
-	SecretAccessKey      string `env:"SECRET_ACCESS_KEY"`
-	UseSSL               bool   `env:"USE_SSL"`
-	MaxRetries           int    `env:"MAX_RETRIES"`
-	PresignedUrlDuration string `env:"PRESIGNED_URL_DURATION"`
-}
-
-func (cfg *S3Config) GetPresignedUrlDuration() (time.Duration, error) {
-	duration, err := time.ParseDuration(cfg.PresignedUrlDuration)
-	if err != nil {
-		return time.Duration(0), err
-	}
-
-	return duration, nil
-}
-
-func (cfg *S3Config) GetOptions() *minio.Options {
-	return &minio.Options{
-		Creds:      credentials.NewStaticV4(cfg.AccessKey, cfg.SecretAccessKey, ""),
-		Secure:     cfg.UseSSL,
-		MaxRetries: cfg.MaxRetries,
-	}
-}
-
-func (cfg *S3Config) GetEndpoint() string {
-	return fmt.Sprintf("%s:%d", cfg.Host, cfg.Port)
 }
 
 type Client struct {
@@ -205,7 +173,7 @@ func (c *Client) BucketExists(ctx context.Context, bucketName string) (bool, err
 	return c.client.BucketExists(ctx, bucketName)
 }
 
-func CreateClient(cfg *S3Config) (IClient, error) {
+func CreateClient(cfg *S3Config) (*Client, error) {
 	client, err := minio.New(cfg.GetEndpoint(), cfg.GetOptions())
 	if err != nil {
 		return nil, err
